+++ src/vic/registers.rs
@@ -3,54 +3,45 @@
 use serde::{Deserialize, Serialize};
 use std::ops::{Index, IndexMut};
 
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct GlobalColors(pub [u8; 3]);
-
-impl GlobalColors {
-    pub const BACKGROUND: u32 = 0;
-    pub const BORDER: u32 = 1;
-    pub const AUX: u32 = 2;
+#[derive(Debug, Serialize, Deserialize, Copy, Clone)]
+pub enum Register {
+    Background,
+    Border,
+    Aux,
 }
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct GlobalColors {
+    pub background: u8,
+    pub border: u8,
+    pub aux: u8,
+}
+
 impl Default for GlobalColors {
     fn default() -> Self {
-        Self([0, 1, 2])
-    }
-}
-impl Index<u32> for GlobalColors {
-    type Output = u8;
-    fn index(&self, index: u32) -> &Self::Output {
-        &self.0[index as usize]
-    }
-}
-impl IndexMut<u32> for GlobalColors {
-    fn index_mut(&mut self, index: u32) -> &mut Self::Output {
-        &mut self.0[index as usize]
-    }
-}
-impl IntoIterator for GlobalColors {
-    type Item = u8;
-    type IntoIter = GlobalColorsIntoIterator;
-
-    fn into_iter(self) -> Self::IntoIter {
-        Self::IntoIter {
-            value: self,
-            index: 0,
+        Self {
+            background: 0,
+            border: 1,
+            aux: 2,
         }
     }
 }
-pub struct GlobalColorsIntoIterator {
-    value: GlobalColors,
-    index: u32,
+impl Index<Register> for GlobalColors {
+    type Output = u8;
+    fn index(&self, index: Register) -> &Self::Output {
+        match index {
+            Register::Background => &self.background,
+            Register::Border => &self.border,
+            Register::Aux => &self.aux,
+        }
+    }
 }
-impl Iterator for GlobalColorsIntoIterator {
-    type Item = u8;
-    fn next(&mut self) -> Option<Self::Item> {
-        if (self.index as usize) < self.value.0.len() {
-            let c = self.value[self.index];
-            self.index += 1;
-            Some(c)
-        } else {
-            None
+impl IndexMut<Register> for GlobalColors {
+    fn index_mut(&mut self, index: Register) -> &mut Self::Output {
+        match index {
+            Register::Background => &mut self.background,
+            Register::Border => &mut self.border,
+            Register::Aux => &mut self.aux,
         }
     }
 }